Reorganize admin module structure
- Move admin-related files from resources to a dedicated admin package
- Update import paths in main.py and test files
- Improve code organization and separation of concerns
