+++ app/admin/__init__.py
@@ -1,0 +1,5 @@
+"""Admin interface web pages for the API."""
+
+from .admin import register_admin
+
+__all__ = ["register_admin"]
